Gen_AI/langchain/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain-aws
langserve
boto3
fastapi 
uvicorn 
openai 
pydantic
sse_starlette